+++ satella/cassandra/common.py
@@ -1,0 +1,5 @@
+try:
+    from cassandra.cluster import ResponseFuture
+except ImportError:
+    class ResponseFuture:
+        pass