+++ tests/test_cassandra.py
@@ -1,8 +1,51 @@
-from satella.cassandra import parallel_for
+from satella.coding.concurrent import CallableGroup
+
+from satella.cassandra import parallel_for, wrap_future
 import unittest
 
 
 class TestCassandra(unittest.TestCase):
+    def test_wrap_future(self):
+        class MockCassandraFuture:
+            def __init__(self):
+                self.value = None
+                self.callbacks = CallableGroup()
+                self.errbacks = CallableGroup()
+
+            def add_callback(self, callback):
+                self.callbacks.add(callback)
+
+            def add_errback(self, errback):
+                self.errbacks.add(errback)
+
+            def set_result(self, x):
+                self.value = x
+                if isinstance(x, Exception):
+                    self.errbacks(x)
+                else:
+                    self.callbacks(x)
+
+        mcf = MockCassandraFuture()
+        wrapped = wrap_future(mcf)
+        a = {}
+
+        def on_done(fut):
+            if fut.exception() is None:
+                a['success'] = True
+            else:
+                a['failure'] = True
+
+        wrapped.add_done_callback(on_done)
+
+        mcf.set_result(None)
+        self.assertTrue(a['success'])
+
+        mcf = MockCassandraFuture()
+        wrapped = wrap_future(mcf)
+        wrapped.add_done_callback(on_done)
+        mcf.set_result(Exception())
+        self.assertTrue(a['failure'])
+
     def test_parallel_for(self):
         class Cursor:
             def __init__(self):