+++ satella/coding/structures/dictionaries/__init__.py
@@ -2,7 +2,8 @@ from .cache_dict import CacheDict
 from .dict_object import apply_dict_object, DictObject
 from .expiring import ExpiringEntryDict, SelfCleaningDefaultDict
 from .objects import DirtyDict, DictionaryView, KeyAwareDefaultDict, TwoWayDictionary
+from .writeback_cache import ExclusiveWritebackCache
 
 __all__ = ['DictObject', 'DirtyDict', 'DictionaryView', 'CacheDict', 'KeyAwareDefaultDict',
            'TwoWayDictionary', 'apply_dict_object', 'ExpiringEntryDict',
-           'SelfCleaningDefaultDict']
+           'SelfCleaningDefaultDict', 'ExclusiveWritebackCache']