+++ satella/coding/concurrent/__init__.py
@@ -8,8 +8,10 @@ from .thread import TerminableThread, Condition, SingleStartThread, call_in_sepa
 from .id_allocator import IDAllocator
 from .timer import Timer
 from .functions import parallel_execute, run_as_future
+from .sync_threadpool import sync_threadpool
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
            'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber',
            'CallNoOftenThan', 'SingleStartThread', 'IDAllocator', 'call_in_separate_thread',
-           'BogusTerminableThread', 'Timer', 'parallel_execute', 'run_as_future']
+           'BogusTerminableThread', 'Timer', 'parallel_execute', 'run_as_future',
+           'sync_threadpool']