+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.16
+
+* added `to_iterator`