+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.11.16
 
 * added `to_iterator`
+* added `class_or_instancemethod`