+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # v2.11.17
+
+* added `wrap_future`
+* refactored `satella.cassandra`
+* deprecated tracing Cassandra's ResponseFutures directly
+