+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.11.18
 
 * added `Condition.notifyAll`
+* added `ExclusiveWritebackCachce`