+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # v2.11.19
+
+* added capability to self-clean to ExclusiveWritebackCache
+* ExclusiveWritebackCache is now thread-safe
+