+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 * added capability to self-clean to ExclusiveWritebackCache
 * ExclusiveWritebackCache is now thread-safe
+* added __slots__ to `ExclusiveWritebackCache`
+* added `store_key_errors` to `ExclusiveWritebackCache`