+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 * ExclusiveWritebackCache is now thread-safe
 * added __slots__ to `ExclusiveWritebackCache`
 * added `store_key_errors` to `ExclusiveWritebackCache`
-
+* added `sync_threadpool`