+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# v2.11.20
+# v2.11.21
+
+Bugfix release
+
 
 * unit testing engine changed from nose2 to pytest, since nose2 couldn't run tests on multiple threads
     It's much faster this way.